Use string for class_year

+++ app/Http/Controllers/UserController.php
@@ -16,7 +16,7 @@ class UserController extends Controller
         $data = $request->validate([
             'expo_token' => ['nullable', 'string'],
             'macpass_number' => ['nullable', 'string', 'numeric', 'digits:9'],
-            'class_year' => ['nullable', 'integer', 'digits:4'],
+            'class_year' => ['nullable', 'numeric', 'digits:4'],
             'position' => ['nullable', 'string', 'in:student,professor,staff'],
             'po_combination' => ['nullable', 'string', 'regex:/^\d+-\d+-\d+$/'],
             'po_number' => ['nullable', 'integer']
